+++ starsky-tools/thumbnail/thumbnail.ts
@@ -1,16 +1,19 @@
 #!/usr/bin/env node
 
-import * as path from 'path';
-import { Query } from './thumbnail-core';
-require('dotenv').config({path:path.join(__dirname,"../", ".env")});
 import { TaskQueue } from 'cwait';
+import * as path from 'path';
+import { Files, Parser, Query } from './thumbnail-core';
+require('dotenv').config({ path: path.join(__dirname, "../", ".env") });
 
 var base_url = process.env.STARKSYBASEURL;
 var access_token = process.env.STARKSYACCESSTOKEN;
 
-if(!base_url || !access_token) {
-	throw new Error ("Missing env's STARKSYBASEURL or STARKSYACCESSTOKEN")
+if (!base_url || !access_token) {
+	throw new Error("Missing env's STARKSYBASEURL or STARKSYACCESSTOKEN")
 }
+
+// Cleanup old files
+new Files().RemoveOldFiles();
 
 function ShowHelpDialog() {
 	console.log("Starksy Remote Thumbnail Helper")
@@ -19,47 +22,52 @@ function ShowHelpDialog() {
 	console.log("use a keyword to search and check if thumbnails are created")
 }
 
-
-function parseArgs() : string[] {
+function parseArgs(): string[] {
 	var args = process.argv.slice(2);
 	if (args.length >= 1) {
-		var parsed = parseInt(args[0])
+		var parsedInt = parseInt(args[0])
+		var isRange = args[0].indexOf("-") >= 1;
+
 		if (args[0] === "-h" || args[0] === "--h" || args[0] === "help") {
 			ShowHelpDialog();
 		}
-		if (isNaN(parsed)) {
+		else if (isRange && args[0].split("-").length === 2) {
+			return new Parser().parseRanges(args);
+		}
+		if (isNaN(parsedInt)) {
 			return [args[0]];
 		}
-		else if(parsed === 0 ){
+		else if (parsedInt === 0) {
 			// Search for today
 			return ["-Datetime>0 -ImageFormat:jpg -!delete"];
 		}
 		else {
 			// 1 = yesterday
-			return ["-Datetime>" + parsed  + " -Datetime<"+ (parsed - 1) + " -ImageFormat:jpg -!delete"];
+			return ["-Datetime>" + parsedInt + " -Datetime<" + (parsedInt - 1) + " -ImageFormat:jpg -!delete"];
 		}
 	}
-
 	var parsedDefault = 1;
-	return ["-Datetime>" + (parsedDefault)  + " -ImageFormat:jpg -!delete"];
+	return ["-Datetime>" + (parsedDefault) + " -ImageFormat:jpg -!delete"];
 }
 
-var searchQueries = parseArgs();
+var query = new Query(base_url, access_token);
 
-var query = new Query(base_url,access_token);
+runQueryChain(0, parseArgs());
 
-searchQueries.forEach(function(searchQuery) {
-	console.log(searchQuery + "\n^^^^searchQuery^^^^");
+function runQueryChain(index = 0, searchQueries: string[]) {
+	if (searchQueries.length === 0) return;
+	if (index >= searchQueries.length) return;
 
-	query.isImportOrDirectSearch(searchQuery).then(async (fileHashList : Array<string>) => {
+	console.log(searchQueries[index] + "\n^^^^searchQuery^^^^");
 
+	query.isImportOrDirectSearch(searchQueries[index]).then(async (fileHashList: Array<string>) => {
 		process.stdout.write("∞ " + fileHashList.length + " ∞");
 
 		// Down chain
 		const queueAxios = new TaskQueue(Promise, query.MAX_SIMULTANEOUS_DOWNLOADS);
 		const axiosResponses = await Promise.all(fileHashList.map(queueAxios.wrap(
-			async (fileHash : string) 	=> 	{
-				if(await query.checkIfSingleFileNeedsToBeDownloaded(fileHash)) {
+			async (fileHash: string) => {
+				if (await query.checkIfSingleFileNeedsToBeDownloaded(fileHash)) {
 					await query.downloadBinarySingleFile(fileHash);
 					return fileHash;
 				}
@@ -67,7 +75,7 @@ searchQueries.forEach(function(searchQuery) {
 		)));
 
 		// Filter before send it to the up chain
-		var filteredAxiosResponses : Array<string> = axiosResponses.filter(function (el) {
+		var filteredAxiosResponses: Array<string> = axiosResponses.filter(function (el) {
 			return el != undefined;
 		});
 
@@ -76,9 +84,9 @@ searchQueries.forEach(function(searchQuery) {
 		// Up chain
 		const queueResizeChain = new TaskQueue(Promise, query.MAX_SIMULTANEOUS_DOWNLOADS);
 		await Promise.all(filteredAxiosResponses.map(queueResizeChain.wrap(
-			async (fileHash : string) 	=> 	{
-				if(await query.resizeImage(fileHash)) {
-					if(await query.uploadTempFile(fileHash)) {
+			async (fileHash: string) => {
+				if (await query.resizeImage(fileHash)) {
+					if (await query.uploadTempFile(fileHash)) {
 						return fileHash; // return isn't working good
 						// resizeChain> [undefined,und..]
 					}
@@ -86,14 +94,18 @@ searchQueries.forEach(function(searchQuery) {
 			}
 		)));
 
-
 		// and clean afterwards
 		query.deleteSourceTempFolder(filteredAxiosResponses);
 		query.deleteTempFolder(filteredAxiosResponses);
 
-		console.log("   `done");
+		console.log("   `done " + index + "/" + (searchQueries.length - 1));
 
-	}).catch( err => {
+		// Next
+		index++;
+		runQueryChain(index, searchQueries);
+
+	}).catch(err => {
 		console.log('err- downloadBinaryApiChain', err);
 	})
-});
+
+}