+++ starsky/starsky/clientapp/src/components/menu-detailview.tsx
@@ -2,11 +2,14 @@
 import { Link } from '@reach/router';
 import React, { useEffect } from 'react';
 import { DetailViewContext } from '../contexts/detailview-context';
+import useKeyboardEvent from '../hooks/use-keyboard-event';
 import useLocation from '../hooks/use-location';
 import { IExifStatus } from '../interfaces/IExifStatus';
 import FetchPost from '../shared/fetch-post';
+import { Keyboard } from '../shared/keyboard';
 import { URLPath } from '../shared/url-path';
 import ModalExport from './modal-export';
+import ModalRenameFile from './modal-rename-file';
 import MoreMenu from './more-menu';
 
 const MenuDetailView: React.FunctionComponent = () => {
@@ -72,13 +75,20 @@ const MenuDetailView: React.FunctionComponent = () => {
     }
   }
 
+  useKeyboardEvent(/(Delete)/, (event: KeyboardEvent) => {
+    if (new Keyboard().isInForm(event)) return;
+    TrashFile();
+  })
+
   var headerName = isDetails ? "header header--main header--edit" : "header header--main";
   if (isMarkedAsDeleted) headerName += " " + "header--deleted"
 
   const [isModalExportOpen, setModalExportOpen] = React.useState(false);
+  const [isModalRenameFileOpen, setModalRenameFileOpen] = React.useState(false);
 
   return (<>
     {isModalExportOpen ? <ModalExport handleExit={() => setModalExportOpen(!isModalExportOpen)} select={[detailView.subPath]} isOpen={isModalExportOpen} /> : null}
+    {isModalRenameFileOpen ? <ModalRenameFile handleExit={() => setModalRenameFileOpen(!isModalRenameFileOpen)} isOpen={isModalRenameFileOpen} /> : null}
 
     <header className={headerName}>
       <div className="wrapper">
@@ -88,7 +98,7 @@ const MenuDetailView: React.FunctionComponent = () => {
           <li className="menu-option" onClick={() => setModalExportOpen(!isModalExportOpen)}>Exporteer</li>
           {!isDetails ? <li className="menu-option" onClick={() => { toggleLabels() }}>Labels</li> : null}
           <li className="menu-option disabled" onClick={() => { alert("werkt nog niet"); }}>Verplaats</li>
-          <li className="menu-option disabled" onClick={() => { alert("werkt nog niet"); }}>Naam wijzigen</li>
+          <li className="menu-option disabled" onClick={() => setModalRenameFileOpen(!isModalRenameFileOpen)}>Naam wijzigen</li>
           <li className="menu-option" onClick={() => { TrashFile(); }}>{!isMarkedAsDeleted ? "Weggooien" : "Undo Weggooien"}</li>
           <li className="menu-option disabled" onClick={() => { alert("werkt nog niet"); }}>Roteer naar rechts</li>
         </MoreMenu>